[RESTEASY-1721] Eliminate ContextParameterInjector's dependence on servlet classes
[RESTEASY-1721] Remove another reference to ServletOutputStream.

[RESTEASY-1721] Load ContextServletOutputStream by reflection.

[RESTEASY-1721] Make variables final in ContextServletOutputStream.

[RESTEASY-1721] Hide more details of the implementation.

[RESTEASY-1721] Checkstyle fix.

[RESTEASY-1723] Hide two more methods in ContextParameterInjector.
